backend/requirements.txt
alembic==1.14.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cv==1.0.0
et-xmlfile==1.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-MySQL==1.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
load-dotenv==0.1.0
Mako==1.3.7
MarkupSafe==3.0.2
msal==1.31.0
mysql-connector-python==9.1.0
nltk==3.9.1
numpy==2.1.2
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pdf2image==1.17.0
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pytesseract==0.3.13
python-dotenv==1.0.1
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
